Improved custom shaders support
Corrected issues with textures loading
